requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
WTForms==3.1.0
Flask-WTF==1.2.1
WTForms-SQLAlchemy==0.3.0

